+++ subot/ui_areas/ui_ocr_types.py
@@ -4,6 +4,9 @@ from subot.ui_areas.CreatureReorderSelectFirst import OCRCreatureRecorderSelectF
 from subot.ui_areas.OCRGodForgeSelect import OCRGodForgeSelectSystem
 from subot.ui_areas.creatures_display import OCRCreaturesDisplaySystem
 from subot.ui_areas.summoning import OcrSummoningSystem
+from subot.ui_areas.OcrUnknownArea import OcrUnknownArea
 
 OCR_UI_SYSTEMS = Union[
-    OCRCreatureRecorderSelectFirst, OCRCreatureRecorderSwapWith, OCRGodForgeSelectSystem, OCRCreaturesDisplaySystem, OcrSummoningSystem]
+    OCRCreatureRecorderSelectFirst, OCRCreatureRecorderSwapWith, OCRGodForgeSelectSystem, OCRCreaturesDisplaySystem, OcrSummoningSystem, OcrUnknownArea]
+
+